# docs/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pullwise API
  description: |
    API for the Pullwise platform - AI-powered code review system.

    ## Authentication

    Most endpoints require authentication via JWT Bearer token. Obtain a token via GitHub OAuth2.

    ## Rate Limiting

    - 100 requests per minute per user
    - 20 requests burst

    ## Versioning

    Current API version: v1.0
  version: 1.0.0
  contact:
    name: Pullwise API Support
    url: https://pullwise.ai
    email: api@pullwise.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.pullwise.ai
    description: Production server
  - url: https://staging-api.pullwise.ai
    description: Staging server
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      description: Retrieve information about the authenticated user
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Logout user
      description: Invalidate the current session
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh JWT token
      description: Obtain a new JWT token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid refresh token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via OAuth2 callback
...
  schemas:
    UserDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          default: Bearer
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/UserDTO'
...
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing JWT token
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
              message:
                type: string
                example: Invalid or expired token
...